Refactored  init.rb file
