+++ app/lib/load_patch/editableby_time_entry_patch.rb
@@ -1,0 +1,20 @@
+module LoadPatch::EditablebyTimeEntryPatch
+	def self.included(base)
+		base.class_eval do
+
+			def editable_by?(usr)
+				# === ERPmine_patch Redmine 5.1 for supervisor edit =====
+				wktime_helper = Object.new.extend(WktimeHelper)
+				if ((!user.blank? && wktime_helper.isSupervisorForUser(user.id)) && wktime_helper.canSupervisorEdit)
+					true
+				else
+				# =============================
+					visible?(usr) && (
+						(usr == user && usr.allowed_to?(:edit_own_time_entries, project)) || usr.allowed_to?(:edit_time_entries, project)
+					)
+				end
+			end
+
+		end
+	end
+end
