+++ app/lib/send_patch/custom_fields_helper_patch.rb
@@ -1,0 +1,5 @@
+module SendPatch::CustomFieldsHelperPatch
+	def self.included(base)
+		CustomFieldsHelper::CUSTOM_FIELDS_TABS << {:name => 'WktimeCustomField', :partial => 'custom_fields/index', :label => :label_wk_time}
+	end
+end
