+++ app/lib/send_patch/issues_controller_patch.rb
@@ -1,4 +1,4 @@
-module IssuesControllerPatch
+module SendPatch::IssuesControllerPatch
 	def self.included(base)
 		base.class_eval do
