+++ app/lib/send_patch/my_controller_patch.rb
@@ -1,0 +1,8 @@
+# Shown users attachment in myaccount page
+module SendPatch::MyControllerPatch
+	def self.included(base)
+		base.class_eval do
+			helper WkdocumentHelper
+	  	end
+	end
+end
