+++ app/lib/send_patch/queries_controller_patch.rb
@@ -1,6 +1,7 @@
-module QueriesControllerPatch
+module SendPatch::QueriesControllerPatch
 	def self.included(base)
 		base.class_eval do
+
 			def redirect_to_wk_expense_entry_query(options)
 			  redirect_to _time_entries_path(@project, nil, options)
 			end
@@ -8,6 +9,7 @@ module QueriesControllerPatch
 			def redirect_to_wk_material_entry_query(options)
 			  redirect_to _time_entries_path(@project, nil, options)
 			end
+
 		end
 	end
 end
